Move body field to UKHSAPage model
